# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stator",
    "version": "1",
    "description": "The wiki's HTTP API. Every endpoint answers JSON; errors share one envelope. Sign in with a session cookie or an API token sent as a bearer token."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This deployment"
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Sign in a local account with email and password, and set the session cookie.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "End the session.",
        "tags": [
          "auth"
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me",
        "summary": "Who is signed in, and the organizations they may act in.",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/callback": {
      "get": {
        "operationId": "oIDCCallback",
        "summary": "Where the identity provider sends the browser back; sets the session cookie.",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found: the browser is sent on."
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/{orgSlug}/start": {
      "get": {
        "operationId": "oIDCStart",
        "summary": "Begin signing in through the organization's identity provider.",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "orgSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "next",
            "in": "query",
            "description": "A path of this application to land on afterwards.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found: the browser is sent on."
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/switch-org": {
      "post": {
        "operationId": "switchOrg",
        "summary": "Move the session to another organization.",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchOrgRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "organization": {
                      "$ref": "#/components/schemas/CurrentOrg"
                    }
                  },
                  "required": [
                    "organization"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/oidc-provider": {
      "get": {
        "operationId": "getOIDCProvider",
        "summary": "The organization's identity provider, if one is configured. For administrators.",
        "tags": [
          "access"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "saveOIDCProvider",
        "summary": "Configure the organization's identity provider. For administrators.",
        "tags": [
          "access"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveOIDCProviderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderView"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "openAPI",
        "summary": "This document.",
        "tags": [
          "health"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/themes": {
      "get": {
        "operationId": "listThemes",
        "summary": "Themes the caller may use: theirs, then the shared ones.",
        "tags": [
          "themes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "themes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Theme"
                      }
                    }
                  },
                  "required": [
                    "themes"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTheme",
        "summary": "Make a theme.",
        "tags": [
          "themes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "theme": {
                      "$ref": "#/components/schemas/Theme"
                    }
                  },
                  "required": [
                    "theme"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/themes/active": {
      "get": {
        "operationId": "activeTheme",
        "summary": "The theme the caller sees: chosen, the organization's default, or null for the built-in one.",
        "tags": [
          "themes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "string"
                    },
                    "theme": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Theme"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "source",
                    "theme"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "chooseTheme",
        "summary": "Use a theme; null returns to the organization's default, null with builtIn keeps the built-in one.",
        "tags": [
          "themes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChooseThemeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "theme": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Theme"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "theme"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/themes/default": {
      "put": {
        "operationId": "setDefaultTheme",
        "summary": "Name the shared theme everybody sees until they choose, or null for the built-in one.",
        "tags": [
          "themes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultThemeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "theme": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/Theme"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "theme"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/themes/examples": {
      "get": {
        "operationId": "themeExamples",
        "summary": "The themes shipped with the product, to start a theme from.",
        "tags": [
          "themes"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "examples": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Example"
                      }
                    }
                  },
                  "required": [
                    "examples"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/themes/import": {
      "post": {
        "operationId": "importTheme",
        "summary": "Make a theme of the caller's own from an exported theme file, sent as a multipart part named file.",
        "tags": [
          "themes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "theme": {
                      "$ref": "#/components/schemas/Theme"
                    }
                  },
                  "required": [
                    "theme"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/themes/{themeID}/assets": {
      "post": {
        "operationId": "uploadThemeAsset",
        "summary": "Put a picture or a font on a theme, as a multipart part named file.",
        "tags": [
          "themes"
        ],
        "parameters": [
          {
            "name": "themeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asset": {
                      "$ref": "#/components/schemas/Asset"
                    }
                  },
                  "required": [
                    "asset"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "An error, in the one shape every endpoint uses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "organization": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CurrentOrg"
              },
              {
                "type": "null"
              }
            ]
          },
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            }
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user",
          "organization",
          "organizations"
        ]
      },
...
      "SwitchOrgRequest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "slug"
        ]
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "stator_session",
        "description": "The cookie a sign-in sets."
      },
      "token": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token."
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "token": []
    }
  ]
}